feat(event): Refactorizar los imports para usar constantes de CheckInStatus y OverlapStatus; actualizar las exportaciones del modelo para incluir Audit
