+++ src/models/index.ts
@@ -1,9 +1,10 @@
 import { Room } from "./room";
 import { Event } from "./event";
 import { User } from "./user";
+import { Audit } from "./audit";
 import { Forecast } from "./forecast"
 import { asociacionesModelos } from "./assoc";
 
 asociacionesModelos();
 
-export { Room, Event, User, Forecast };
+export { Room, Event, User, Forecast, Audit };
